coretests/num: use ldexp instead of hard-coding a power of 2
